refactor: moved request_schemas.py and response_schemas.py to their own package
