+++ backend/README.md
@@ -1,3 +1,9 @@
 # Deployment
 
-Deploy the backend using `uvicorn server:app --reload --port 8001`
+Deploy the backend using `uvicorn backend.server:app --reload --port 8001`. This should be run from the main directory,
+not the `/backend` directory.
+
+# CLI Usage
+
+Run the CLI using `python -m backend.client <options>`. This should be run from the main directory, not the `\backend`
+directory.
