+++ models/response_schemas.py
@@ -5,6 +5,9 @@ from pydantic import BaseModel
 
 
 class ItemResponse(BaseModel):
+    """
+    Model representing an item returned by the /item endpoint.
+    """
     name: str
     unit_weight: int
     price: int
@@ -13,11 +16,17 @@ class ItemResponse(BaseModel):
 
 
 class TransactionItemResponse(BaseModel):
+    """
+    Model representing a single item transaction (name, quantity pair)
+    """
     item_name: str
     item_quantity: int
 
 
 class TransactionResponse(BaseModel):
+    """
+    Model representing a transaction involving multiple item transactions (name, quantity pairs)
+    """
     transaction_id: int
     student_id: Optional[str]
     day_of_week: str
@@ -27,6 +36,9 @@ class TransactionResponse(BaseModel):
 
 
 class MessageResponse(BaseModel):
+    """
+    Model representing a message sent by the server
+    """
     message: str
